Some more gltform changes
In this PR we make some more gltform changes, in light of some hpegl
testing.  At present WriteGLConfig writes the .gltform file in home and
in the cwd.  GetGLConfig then looks for the .gltform file first in home
and then in the cwd.  The issue is that if you run hpegl with a bmaas
stanza and then subsequently without a bmaas stanza then hpegl will try
to initialise bmaas because it finds a .gltform in home.  This is not very
nice from a usability perspective.  This PR changes things so that
WriteGLConfig writes a .gltform only in the cwd.  Also GetGLConfig looks
first in cwd and then in home.  This means that what we expect to be the
most common use-case of a user specifying the .gltform contents in the
hpegl bmaas stanza will behave in a "friendly" way with no unexpected
attempts to initialise bmaas.