Remove internal IP address from token POST in k8s environments
When running the hpegl provider in k8s environments a bogus IAM Issuer
URL in the supplied API client creds will result in the internal IP
address of the running container being contained in the error message
from the IAM POST command to retrieve a token.  This PR sanitises the
relevant error message and removes the IP address.

Signed-off-by: Eamonn O'Toole <eamonn.otoole@hpe.com>