Some go.mod updates
In this PR we update the dependencies in go.mod.  We advance the
go.mod version to 1.18.  We change ci.yaml to use go versions 1.19
and 1.18.

We change the Makefile tools rule to install a specific version of
golangci-lint.

We had to make some minor changes as a result of linter errors.

Signed-off-by: Eamonn O'Toole <eamonn.otoole@hpe.com>