Changes to support GLP IAM API Client
In this PR we add support for GLP IAM API Clients.  We retain support
for GLCS API Clients.  We also retain support for "non-api vended" GLCS
API Clients for backwards compatibility.

We've added a new provider parameter "iam_version" with the
corresponding env-var HPEGL_IAM_VERSION to designate which version of
IAM is being used:
- glcs
- glp

The default value of this parameter is glcs.  We've added a function to
validate the value of "iam_version".

We've updated the "description" field for "api_vended_service_client"
and "tenant_id" in light of the support for GLP IAM.

The issuertoken package has been modified to generate the correct URL
and the correct set of http request parameters depending on which type
of IAM is being used.

We've had to update various unit-tests and the generated mocks.

Signed-off-by: Eamonn O'Toole <eamonn.otoole@hpe.com>