Add support for GLP IAM role and workspace to gltform
The Metal team have added support for GLP IAM to their service.  This
means that glp_role and glp_workspace must be specified in the metal
block in the provider stanza.  This PR adds the following two entries to
gltform
- glp_workspace
- glp_role

These entries will be written-out during Metal client initialisation in
terraform-provider-hpegl.

Signed-off-by: Eamonn O'Toole <eamonn.otoole@hpe.com>