Create new http request on each retry
In this PR we change the behaviour in issuertoken to:
- create a new http request object on each retry
- set Close on each request so that the underlying tcp connection is
  always closed

We do this because we are seeing issues with retries on 502 responses
from IAM.  Retries don't seem to make any difference, in other words we
continue to receive 502 responses on each retry.

We fix unit-tests.

We also log each retry so that the terraform provider logs will record
the retries.

Signed-off-by: Eamonn O'Toole <eamonn.otoole@hpe.com>